factor historical geography controller functions into helper
